Added Numpad to end the Game
We still need to add textures to the keys, plus making a "real" end (the
door opens and the camera turns towards it). We might also add a sound
effect when the code is wrong.

The testing has been done though, as I used Debug.Log to check whether
the functions were correctly called.