Added "HeadTracking"
I translated the whole scene at (0,0,0) otherwise it won't work.
Moreover, if we want to "move" around the scene, it's "MoveTheScene"
(containing all assets) that needs to be moved. This would give a
semblance of movement without messing up the headtracking that bugs if
we move the camera (even moving both of them out of (0,0,0) makes it
bug).

Thus we need to change the script "MoveAroundScene.cs" so that it moves
in the other direction (it should be only a - to be added) and attach it
to "MoveTheScene" instead of "MainCamera".